requirements.txt
streamlit>=1.30
st-diff-viewer
google-genai
requests
python-dotenv
